Add images (see description)
Originals are `a` and `b`, extracted from PDFs. The others are edits of those for testing robustness with more realistic color and noise.
Also added ground truths for automated comparison of images.
Maximum dimension of any image is 500px. 
`a` and `b` sets are all the same dimensions relative to the respective originals. `a` and `b` originals had different aspect ratios when exported, and so are different here.